+++ packages/gatsby-transformer-remark/src/__tests__/extend-node.js
@@ -33,7 +33,7 @@ async function queryResult(
         get: () => null,
         set: () => null,
       },
-      getNodesByType: type => [],
+      getNodesByType: (type) => [],
       ...additionalParameters,
     },
     {
@@ -100,15 +100,15 @@ const bootstrapTest = (
     },
   }
   // Make some fake functions its expecting.
-  const loadNodeContent = node => Promise.resolve(node.content)
+  const loadNodeContent = (node) => Promise.resolve(node.content)
 
-  it(label, async done => {
+  it(label, async (done) => {
     node.content = content
-    const createNode = markdownNode => {
+    const createNode = (markdownNode) => {
       queryResult([markdownNode], query, {
         additionalParameters,
         pluginOptions,
-      }).then(result => {
+      }).then((result) => {
         if (result.errors) {
           done.fail(result.errors)
         }
@@ -156,7 +156,7 @@ Where oh where is my little pony?`,
           title
       }
       `,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.excerpt).toBe(`Where oh where is my little pony?`)
       expect(node.excerptAst).toMatchObject({
@@ -191,7 +191,7 @@ date: "2017-09-18T23:19:51.246Z"
           title
       }
       `,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.excerpt).toBe(``)
       expect(node.excerptAst).toMatchObject({
@@ -220,7 +220,7 @@ In quis lectus sed eros efficitur luctus. Morbi tempor, nisl eget feugiat tincid
           title
       }
       `,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.excerpt).toBe(`Where oh where is my little pony?`)
       expect(node.excerptAst).toMatchObject({
@@ -263,7 +263,7 @@ In quis lectus sed eros efficitur luctus. Morbi tempor, nisl eget feugiat tincid
     `given PLAIN correctly uses excerpt separator`,
     contentWithSeparator,
     `excerpt(format: PLAIN)`,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.excerpt).toBe(`Where oh where is my little pony?`)
     },
@@ -274,7 +274,7 @@ In quis lectus sed eros efficitur luctus. Morbi tempor, nisl eget feugiat tincid
     `given HTML correctly uses excerpt separator`,
     contentWithSeparator,
     `excerpt(format: HTML)`,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.excerpt).toBe(
         `<p>Where oh where <strong>is</strong> my little pony?</p>\n`
@@ -287,7 +287,7 @@ In quis lectus sed eros efficitur luctus. Morbi tempor, nisl eget feugiat tincid
     `given MARKDOWN correctly uses excerpt separator`,
     contentWithSeparator,
     `excerpt(format: MARKDOWN)`,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.excerpt).toBe(`Where oh where **is** my little pony?\n`)
     },
@@ -307,7 +307,7 @@ In quis lectus sed eros efficitur luctus. Morbi tempor, nisl eget feugiat tincid
     `given MARKDOWN without excerpt separator, falls back to pruneLength`,
     contentWithoutSeparator,
     `excerpt(pruneLength: 40, format: MARKDOWN)`,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.excerpt.length).toBe(45)
       expect(node.excerpt).toBe(
@@ -321,7 +321,7 @@ In quis lectus sed eros efficitur luctus. Morbi tempor, nisl eget feugiat tincid
     `given MARKDOWN, pruning is done not counting markdown characters`,
     contentWithoutSeparator,
     `excerpt(pruneLength: 19, format: MARKDOWN)`,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       // we want the pruning to preserve markdown chars and not count them in the length
       expect(node.excerpt.length).toBe(23)
@@ -345,7 +345,7 @@ In quis lectus sed eros efficitur luctus. Morbi tempor, nisl eget feugiat tincid
           title
       }
       `,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.excerpt.length).toBe(139)
       expect(node.excerptAst.children.length).toBe(1)
@@ -363,7 +363,7 @@ In quis lectus sed eros efficitur luctus. Morbi tempor, nisl eget feugiat tincid
           title
       }
       `,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.excerpt.length).toBe(46)
       expect(node.excerptAst.children.length).toBe(1)
@@ -381,7 +381,7 @@ In quis lectus sed eros efficitur luctus. Morbi tempor, nisl eget feugiat tincid
           title
       }
       `,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.excerpt.length).toBe(50)
       expect(node.excerptAst.children.length).toBe(1)
@@ -400,7 +400,7 @@ In quis lectus sed eros efficitur luctus. Morbi tempor, nisl eget feugiat tincid
             title
         }
         `,
-      node => {
+      (node) => {
         expect(node).toMatchSnapshot()
         expect(node.excerpt.length).toBe(139)
         expect(node.excerptAst.children.length).toBe(1)
@@ -419,7 +419,7 @@ In quis lectus sed eros efficitur luctus. Morbi tempor, nisl eget feugiat tincid
             title
         }
         `,
-      node => {
+      (node) => {
         expect(node).toMatchSnapshot()
         expect(node.excerpt.length).toBe(46)
         expect(node.excerptAst.children.length).toBe(1)
@@ -438,7 +438,7 @@ In quis lectus sed eros efficitur luctus. Morbi tempor, nisl eget feugiat tincid
             title
         }
         `,
-      node => {
+      (node) => {
         expect(node).toMatchSnapshot()
         expect(node.excerpt.length).toBe(50)
         expect(node.excerptAst.children.length).toBe(1)
@@ -463,7 +463,7 @@ Where oh [*where*](nick.com) **_is_** ![that pony](pony.png)?`,
           title
       }
       `,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.excerpt).toBe(
         `<p>Where oh <a href="nick.com"><em>where</em></a> <strong><em>is</em></strong> <img src="pony.png" alt="that pony">?</p>`
@@ -560,7 +560,7 @@ Where oh [*where*](nick.com) **_is_** ![that pony](pony.png)?`,
           title
       }
       `,
-    node => {
+    (node) => {
       expect(node.excerpt).toBe(`Where oh where is that pony?`)
     },
     {}
@@ -575,7 +575,7 @@ date: "2017-09-18T23:19:51.246Z"
 
  My pony likes space on the left and right! `,
     `excerpt`,
-    node => {
+    (node) => {
       expect(node.excerpt).toBe(`My pony likes space on the left and right!`)
     },
     {}
@@ -592,7 +592,7 @@ My pony is little.
 
 Little is my pony.`,
     `excerpt`,
-    node => {
+    (node) => {
       expect(node.excerpt).toBe(`My pony is little. Little is my pony.`)
     },
     {}
@@ -611,7 +611,7 @@ date: "2017-09-18T23:19:51.246Z"
 
 It's pony time.`,
     `excerpt`,
-    node => {
+    (node) => {
       expect(node.excerpt).toBe(
         `Ponies: The Definitive Guide What time is it? It's pony time.`
       )
@@ -630,7 +630,7 @@ date: "2017-09-18T23:19:51.246Z"
 | -------------- | -------- |
 | My Little Pony | Me, Duh  |`,
     `excerpt`,
-    node => {
+    (node) => {
       expect(node.excerpt).toBe(`Pony Owner My Little Pony Me, Duh`)
     },
     {}
@@ -647,7 +647,7 @@ If my pony ain't broke,${`  `}
 don't fix it.`,
     // ^ Explicit syntax for trailing spaces to not get accidentally trimmed.
     `excerpt`,
-    node => {
+    (node) => {
       expect(node.excerpt).toBe(`If my pony ain't broke, don't fix it.`)
     },
     {}
@@ -666,7 +666,7 @@ date: "2017-09-18T23:19:51.246Z"
 
 Pony express had nothing on my little pony.`,
     `excerpt`,
-    node => {
+    (node) => {
       expect(node.excerpt).toBe(
         `Pony express Pony express had nothing on my little pony.`
       )
@@ -688,7 +688,7 @@ Where is my <code>pony</code> named leo?`,
           title
       }
       `,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.excerpt).toBe(
         `<p>Where is my <code>pony</code> named leo?</p>`
@@ -743,7 +743,7 @@ Where oh where is that pony? Is he in the stable or down by the stream?`,
           title
       }
       `,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.excerpt).toBe(
         `<p>Where oh where is that pony? Is he in the stable…</p>`
@@ -786,7 +786,7 @@ Lorem ipsum dolor sit amet, consectetur adipiscing elit. Morbi auctor sit amet v
         title
     }
     `,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.excerpt).toBe(
         `<p>Where oh where is that <em>pony</em>? Is he in the stable or by the stream?</p>\n`
@@ -851,7 +851,7 @@ In quis lectus sed eros efficitur luctus. Morbi tempor, nisl eget feugiat tincid
     frontmatter {
         title
     }`,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.wordCount).toEqual({
         paragraphs: 2,
@@ -871,7 +871,7 @@ In quis lectus sed eros efficitur luctus. Morbi tempor, nisl eget feugiat tincid
 コンポーネントベース 自分自身の状態を管理するカプセル化されたコンポーネントをまず作成し、これらを組み合わせることで複雑なユーザインターフェイスを構築します。 コンポーネントのロジックは、Template ではなく JavaScript そのもので書くことができるので、様々なデータをアプリケーション内で簡単に取り回すことができ、かつ DOM に状態を持たせないようにすることができます。
 `,
     `timeToRead`,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.timeToRead).toEqual(1)
     }
@@ -894,7 +894,7 @@ date: "2017-09-18T23:19:51.246Z"
     frontmatter {
         title
     }`,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.wordCount).toEqual({
         paragraphs: null,
@@ -911,10 +911,24 @@ date: "2017-09-18T23:19:51.246Z"
     frontmatter {
         title
     }`,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.timeToRead).toBe(1)
     }
+  )
+
+  bootstrapTest(
+    `correctly uses a custom value for timeToRead`,
+    `${content}\nWhere oh where is my little pony?\n`,
+    `timeToRead
+    frontmatter {
+        title
+    }`,
+    (node) => {
+      expect(node).toMatchSnapshot()
+      expect(node.timeToRead).toBe(9)
+    },
+    { pluginOptions: { timeToRead: (wordCount) => wordCount / 1 } }
   )
 })
 
@@ -937,7 +951,7 @@ some other text
     frontmatter {
         title
     }`,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(console.warn).toBeCalled()
     }
@@ -965,7 +979,7 @@ final text
     frontmatter {
         title
     }`,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
     }
   )
@@ -992,7 +1006,7 @@ final text
     frontmatter {
         title
     }`,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
     }
   )
@@ -1014,7 +1028,7 @@ some other text`,
     frontmatter {
         title
     }`,
-    node => {
+    (node) => {
       expect(node.tableOfContents).toBe(`<ul>
 <li><a href="/my%20little%20pony/#first-title">first title</a></li>
 </ul>`)
@@ -1038,7 +1052,7 @@ some other text`,
     frontmatter {
         title
     }`,
-    node => {
+    (node) => {
       expect(node.tableOfContents).toBe(`<ul>
 <li><a href="/my%20little%20pony/#first-title">first title</a></li>
 </ul>`)
@@ -1073,7 +1087,7 @@ final text`,
     frontmatter {
         title
     }`,
-    node => {
+    (node) => {
       expect(node.tableOfContents).toBe(`<ul>
 <li><a href="/my%20little%20pony/#third-title">third title</a></li>
 </ul>`)
@@ -1103,7 +1117,7 @@ It's easier than you may imagine`,
     frontmatter {
         title
     }`,
-    node => expect(node).toMatchSnapshot()
+    (node) => expect(node).toMatchSnapshot()
   )
 
   bootstrapTest(
@@ -1127,7 +1141,7 @@ It's easier than you may imagine`,
     frontmatter {
         title
     }`,
-    node => expect(node).toMatchSnapshot(),
+    (node) => expect(node).toMatchSnapshot(),
     {
       pluginOptions: {
         plugins: [
@@ -1161,7 +1175,7 @@ This is [a reference]
 [a reference]: ./path/to/page2
 `,
     `html`,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.html).toMatch(`<a href="path/to/page1">`)
       expect(node.html).toMatch(`<a href="./path/to/page2">`)
@@ -1185,7 +1199,7 @@ This is [a reference]
 [a reference]: /path/to/page2
 `,
     `html`,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.html).toMatch(`<a href="/prefix/path/to/page1">`)
       expect(node.html).toMatch(`<a href="/prefix/path/to/page2">`)
@@ -1209,7 +1223,7 @@ This is [a reference]
 [a reference]: /path/to/page2
 `,
     `html`,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.html).toMatch(`<a href="/prefix/path/to/page1">`)
       expect(node.html).toMatch(`<a href="/prefix/path/to/page2">`)
@@ -1227,7 +1241,7 @@ console.log('hello world')
 \`\`\`
 `,
     `htmlAst`,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(
         node.htmlAst.children[0].children[0].properties.className
@@ -1251,7 +1265,7 @@ describe(`Headings are generated correctly from schema`, () => {
       value
       depth
     }`,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.headings).toEqual([
         {
@@ -1277,7 +1291,7 @@ describe(`Headings are generated correctly from schema`, () => {
       value
       depth
     }`,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.headings).toEqual([
         {
@@ -1301,7 +1315,7 @@ describe(`Headings are generated correctly from schema`, () => {
       value
       depth
     }`,
-    node => {
+    (node) => {
       expect(node).toMatchSnapshot()
       expect(node.headings).toEqual([
         {