+++ packages/gatsby/src/redux/reducers/index.js
@@ -1,5 +1,9 @@
 const reduxNodes = require(`./nodes`)
 const lokiNodes = require(`../../db/loki/nodes`).reducer
+import { redirectsReducer } from "./redirects"
+import { staticQueryComponentsReducer } from "./static-query-components"
+import { webpackCompilationHashReducer } from "./webpack-compilation-hash"
+import { reducer as logReducer } from "gatsby-cli/lib/reporter/redux/reducer"
 
 const backend = process.env.GATSBY_DB_NODES || `redux`
 
@@ -54,16 +58,16 @@ module.exports = {
   status: require(`./status`),
   componentDataDependencies: require(`./component-data-dependencies`),
   components: require(`./components`),
-  staticQueryComponents: require(`./static-query-components`),
+  staticQueryComponents: staticQueryComponentsReducer,
   jobs: require(`./jobs`),
   jobsV2: require(`./jobsv2`),
   webpack: require(`./webpack`),
-  webpackCompilationHash: require(`./webpack-compilation-hash`),
-  redirects: require(`./redirects`),
+  webpackCompilationHash: webpackCompilationHashReducer,
+  redirects: redirectsReducer,
   babelrc: require(`./babelrc`),
   schemaCustomization: require(`./schema-customization`),
   themes: require(`./themes`),
-  logs: require(`gatsby-cli/lib/reporter/redux/reducer`),
+  logs: logReducer,
   inferenceMetadata: require(`./inference-metadata`),
   pageDataStats: require(`./page-data-stats`),
   pageData: require(`./page-data`),